qwwerr

+++ src/main/java/com/safee/DAO/interfaces/EmployeeDaoInterface.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.safee.model.Employee;
 
 public interface EmployeeDaoInterface {
-	Employee getEmployee(Long employeeNo);
+	Employee getEmployee(Integer employeeNo);
 	void updateEmployee(Employee employee);
 	void addEmployee(Employee employee);
 	public List<Employee> getAllEmployee();
